+++ coral/apps.py
@@ -6,6 +6,9 @@ class CoralConfig(AppConfig):
     is_arches_application = True
 
     def ready(self):
+        from coral.utils.deferred_tile_index import patch_tile_index
+
+        patch_tile_index()
         _patch_arches_search_indexing_factory()